% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_df.R
\name{read.df}
\alias{read.df}
\title{Function to read in data frames with strings as factors = FALSE}
\usage{
read.df(filename, ...)
}
\arguments{
\item{filename}{The name of the file}

\item{...}{Other function calls}
}
\value{
A data.frame containing the contents of filename
}
\description{
Function to read in data frames with strings as factors = FALSE
}
\examples{
df<-read.df(system.file("extdata", "exdf.txt", package = "spfTools"),sep='\t',header=TRUE)
}
