% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem.R
\name{sem}
\alias{sem}
\title{Function to calculate the standard error of the mean}
\usage{
sem(x)
}
\arguments{
\item{x}{vector of values}
}
\value{
sem A numeric that is the SEM
}
\description{
Function to calculate the standard error of the mean
}
\examples{
se<-sem(sample(1:100,10))
}
