% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTools.R
\name{violin_plot}
\alias{violin_plot}
\title{Create violin plots with}
\usage{
violin_plot(x, ..., range = 1.5, h = NULL, ylim = NULL,
  names = NULL, horizontal = FALSE, col = "magenta",
  border = "black", lty = 1, lwd = 1, rectCol = "black",
  colMed = "white", pchMed = 19, at, add = FALSE, wex = 1,
  drawRect = TRUE, plot.axes = TRUE, axis.box = FALSE,
  plot.ann = TRUE)
}
\arguments{
\item{x}{Data to plot}

\item{...}{legend parameters}

\item{range}{Range}

\item{h}{h}

\item{ylim}{y-axis limits}

\item{names}{Axis names}

\item{horizontal}{Plot horizontally or vertically}

\item{col}{Color for the violin backgrounds}

\item{border}{Color for the violin borders}

\item{lty}{Line type}

\item{lwd}{Line width}

\item{rectCol}{Rectangle color}

\item{colMed}{color med}

\item{pchMed}{shape for med}

\item{at}{plot param at}

\item{add}{Add the plot or not}

\item{wex}{wex}

\item{drawRect}{Add a rectangle to the figure}

\item{plot.axes}{Whether or not to plot axes}

\item{axis.box}{Add the axis box}

\item{plot.ann}{Add the annotations to the plot}
}
\value{
Optionally returns the summary statistics of the data.
}
\description{
Create violin plots with
}
\note{
Modified from vioplot() in library(vioplot)
}
\examples{
mu<-2
si<-0.6
bimodal<-c(rnorm(1000,-mu,si),rnorm(1000,mu,si))
uniform<-runif(2000,-4,4)
normal<-rnorm(2000,0,3)
gwsca.vioplot(bimodal,uniform,normal,col=c("red","blue","green"))
}
